+++ BusinessAccountantService/Managers/ExpensesManager.cs
@@ -3,6 +3,7 @@ using BusinessAccountantService.Models;
 using Microsoft.Data.Sqlite;
 using System;
 using System.Collections.Generic;
+using System.Data;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -32,7 +33,7 @@ namespace BusinessAccountantService.Managers
                         stats.Add(new Expense // Создаем объект класса
                         {
                             Category = reader.IsDBNull(0) ? "Прочее" : reader.GetString(0),
-                            Amount = reader.GetDouble(1)
+                            Amount = reader.GetDecimal(1)
                         });
                     }
                 }
@@ -63,7 +64,7 @@ namespace BusinessAccountantService.Managers
                         {
                             Id = reader.GetInt32(0),
                             Description = reader.GetString(1),
-                            Amount = reader.GetDouble(2),
+                            Amount = reader.GetDecimal(2),
                             Category = reader.IsDBNull(3) ? "Прочее" : reader.GetString(3),
                             Date = reader.GetDateTime(4)
                         });
@@ -96,7 +97,7 @@ namespace BusinessAccountantService.Managers
                         stats.Add(new Expense
                         {
                             Category = reader.IsDBNull(0) ? "Прочее" : reader.GetString(0),
-                            Amount = reader.GetDouble(1)
+                            Amount = reader.GetDecimal(1)
                         });
                     }
                 }
